backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
chromadb>=1.0.0
langchain-text-splitters>=0.3.0
pypdf>=5.0.0
python-docx>=1.1.0
python-multipart>=0.0.18
httpx>=0.28.0
aiosqlite>=0.21.0
